Javscript: fix bug with embed parameter
